+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Version 0.4
 -----------
 
+* Rename the project from ``aiogreen`` to ``aioeventlet``
 * Rename the ``link_future()`` function to :func:`yield_future`
 
 2014-10-23: version 0.3